Separated code for SocketProvider
fixed eslint react-refresh error
